feat: Add renderOutdatedCard function to templates.js for displaying outdated post warnings; update main.js to utilize the new function
